requirements.txt
google-genai
flask
pydantic
python-dotenv